GH-47973: [C++][Parquet] Fix invalid Parquet files written when dictionary encoded pages are large (#47998)
### Rationale for this change

Prevents silently writing invalid data when using dictionary encoding and the number of bits in the estimated max buffer size is greater than the max int32 value.

Also fixes an overflow resulting in a "Negative buffer resize" error if the buffer size in bytes is greater than max int32, and instead throw a more helpful exception.

### What changes are included in this PR?

* Fix overflow when computing the bit position in `BitWriter::PutValue`. This overflow would cause the method to return without writing data, and the return value is only checked in debug builds.
* Change buffer size calculations to use int64 and check for overflow before casting to int

### Are these changes tested?

Yes, I've added unit tests for both issues. These require enabling `ARROW_LARGE_MEMORY_TESTS` as they allocate a lot of memory.

### Are there any user-facing changes?

**This PR contains a "Critical Fix".**

This fixes a bug where invalid Parquet files can be silently written when the buffer size for dictionary indices is large.

* GitHub Issue: #47973

Authored-by: Adam Reeve <adreeve@gmail.com>

+++ cpp/src/arrow/util/bit_stream_utils_internal.h
@@ -58,7 +58,7 @@ class BitWriter {
   int buffer_len() const { return max_bytes_; }
 
   /// Writes a value to buffered_values_, flushing to buffer_ if necessary.  This is bit
-  /// packed.  Returns false if there was not enough space. num_bits must be <= 32.
+  /// packed.  Returns false if there was not enough space. num_bits must be <= 64.
   bool PutValue(uint64_t v, int num_bits);
 
   /// Writes v to the next aligned byte using num_bytes. If T is larger than
@@ -197,7 +197,9 @@ inline bool BitWriter::PutValue(uint64_t v, int num_bits) {
     ARROW_DCHECK_EQ(v >> num_bits, 0) << "v = " << v << ", num_bits = " << num_bits;
   }
 
-  if (ARROW_PREDICT_FALSE(byte_offset_ * 8 + bit_offset_ + num_bits > max_bytes_ * 8))
+  if (ARROW_PREDICT_FALSE(static_cast<int64_t>(byte_offset_) * 8 + bit_offset_ +
+                              num_bits >
+                          static_cast<int64_t>(max_bytes_) * 8))
     return false;
 
   buffered_values_ |= v << bit_offset_;
